@feat: 按 tile 图层循环渲染 + 新增物品/陷阱/合成/种植系统
- 重构渲染管线：从按类别批量绘制改为逐 tile 图层循环（地形→物品→生物→覆盖层），修正遮挡关系
- TileMap 新增 TileLayerRenderer 回调接口，GameScene 实现图层委托
- GroundItemManager 添加空间索引，getItemsAt/hasItemAt 从 O(N) 优化为 O(1)
- CreatureGrid/CreatureManager 新增 getAllAtTile/getAliveCreaturesAt 空间查询
- 新增陷阱系统（TrapManager/PlacedTrap）、合成界面（CraftingScreen）
- 新增种植/采集/狩猎等物品动作和配方
- 新增石制工具、农作物、食谱等物品数据
@

+++ src/main/java/com/github/game/cdda/screen/hud/CharacterInfoPanel.java
@@ -141,8 +141,8 @@ public class CharacterInfoPanel implements StatusPanel {
     }
 
     /** 设置速度百分比 */
-    public void setSpeed(int percent) {
-        this.speedPercent = percent;
+    public void setSpeed(long percent) {
+        this.speedPercent = (int) percent;
     }
 
     /**
